refactor: move navigation to app-client
