+++ frontend/apps/app-client/src/components/navigation/MobileToolbar.jsx
@@ -8,7 +8,7 @@ import ListIcon from '@mui/icons-material/List';
 import CloseIcon from '@mui/icons-material/Close';
 import LogoutIcon from '@mui/icons-material/Logout';
 
-import Logo from '../../atoms/Logo';
+import Logo from '../../../../../shared/atoms/Logo';
 
 const StyledMenuItem = styled(MenuItem)(({
 	padding: '6px 3px',
